Call youtube-dl with --no-mtime
Since videos should be downloaded as they become available anyways.
Implemented this because some videos from the creators I watch had wrong
timestamps on the files which caused them to be buried under newer
videos (maybe they were initially uploading the video as
private/unlisted or something and then making it public at a later date,
not sure but this fixes it).